+++ execution_engine/src/core/engine_state/engine_config.rs
@@ -12,6 +12,8 @@ pub const DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT: u32 = 12;
 pub const DEFAULT_MAX_STORED_VALUE_SIZE: u32 = 8 * 1024 * 1024;
 /// Default value for maximum delegators per validator.
 pub const DEFAULT_MAX_DELEGATOR_SIZE_LIMIT: u32 = 950;
+/// Default value for minimum delegation amount in motes.
+pub const DEFAULT_MINIMUM_DELEGATION_AMOUNT: u64 = 500 * 1_000_000_000;
 
 /// The runtime configuration of the execution engine
 #[derive(Debug, Copy, Clone)]
@@ -25,6 +27,7 @@ pub struct EngineConfig {
     max_runtime_call_stack_height: u32,
     max_stored_value_size: u32,
     max_delegator_size_limit: u32,
+    minimum_delegation_amount: u64,
     wasm_config: WasmConfig,
     system_config: SystemConfig,
 }
@@ -37,6 +40,7 @@ impl Default for EngineConfig {
             max_runtime_call_stack_height: DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
             max_stored_value_size: DEFAULT_MAX_STORED_VALUE_SIZE,
             max_delegator_size_limit: DEFAULT_MAX_DELEGATOR_SIZE_LIMIT,
+            minimum_delegation_amount: DEFAULT_MINIMUM_DELEGATION_AMOUNT,
             wasm_config: WasmConfig::default(),
             system_config: SystemConfig::default(),
         }
@@ -45,12 +49,14 @@ impl Default for EngineConfig {
 
 impl EngineConfig {
     /// Creates a new engine configuration with provided parameters.
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         max_query_depth: u64,
         max_associated_keys: u32,
         max_runtime_call_stack_height: u32,
         max_stored_value_size: u32,
         max_delegator_size_limit: u32,
+        minimum_delegation_amount: u64,
         wasm_config: WasmConfig,
         system_config: SystemConfig,
     ) -> EngineConfig {
@@ -60,6 +66,7 @@ impl EngineConfig {
             max_runtime_call_stack_height,
             max_stored_value_size,
             max_delegator_size_limit,
+            minimum_delegation_amount,
             wasm_config,
             system_config,
         }
@@ -93,5 +100,10 @@ impl EngineConfig {
     /// Returns the current system config.
     pub fn system_config(&self) -> &SystemConfig {
         &self.system_config
+    }
+
+    /// Returns the minimum delegation amount in motes.
+    pub fn minimum_delegation_amount(&self) -> u64 {
+        self.minimum_delegation_amount
     }
 }