+++ execution_engine_testing/tests/src/test/regression/gh_2605.rs
@@ -11,8 +11,9 @@ use casper_engine_test_support::{
 use casper_execution_engine::{
     core::{
         engine_state::{
-            engine_config::DEFAULT_MAX_DELEGATOR_SIZE_LIMIT, EngineConfig, Error, UpgradeConfig,
-            DEFAULT_MAX_QUERY_DEPTH, DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
+            engine_config::{DEFAULT_MAX_DELEGATOR_SIZE_LIMIT, DEFAULT_MINIMUM_DELEGATION_AMOUNT},
+            EngineConfig, Error, UpgradeConfig, DEFAULT_MAX_QUERY_DEPTH,
+            DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
         },
         execution,
     },
@@ -301,6 +302,7 @@ fn make_engine_config(max_memory: u32) -> EngineConfig {
         DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
         DEFAULT_MAX_STORED_VALUE_SIZE,
         DEFAULT_MAX_DELEGATOR_SIZE_LIMIT,
+        DEFAULT_MINIMUM_DELEGATION_AMOUNT,
         wasm_config,
         SystemConfig::default(),
     )