+++ execution_engine_testing/tests/src/test/system_costs.rs
@@ -9,8 +9,10 @@ use casper_engine_test_support::{
 };
 use casper_execution_engine::{
     core::engine_state::{
-        engine_config::DEFAULT_MAX_DELEGATOR_SIZE_LIMIT, genesis::GenesisValidator, EngineConfig,
-        GenesisAccount, DEFAULT_MAX_QUERY_DEPTH, DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
+        engine_config::{DEFAULT_MAX_DELEGATOR_SIZE_LIMIT, DEFAULT_MINIMUM_DELEGATION_AMOUNT},
+        genesis::GenesisValidator,
+        EngineConfig, GenesisAccount, DEFAULT_MAX_QUERY_DEPTH,
+        DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
     },
     shared::{
         host_function_costs::{Cost, HostFunction, HostFunctionCosts},
@@ -55,7 +57,7 @@ static VALIDATOR_1: Lazy<PublicKey> = Lazy::new(|| PublicKey::from(&*VALIDATOR_1
 static VALIDATOR_1_ADDR: Lazy<AccountHash> = Lazy::new(|| AccountHash::from(&*VALIDATOR_1));
 const VALIDATOR_1_STAKE: u64 = 250_000;
 const BOND_AMOUNT: u64 = 42;
-const BID_AMOUNT: u64 = 99;
+const BID_AMOUNT: u64 = 99 + DEFAULT_MINIMUM_DELEGATION_AMOUNT;
 const TRANSFER_AMOUNT: u64 = 123;
 const BID_DELEGATION_RATE: DelegationRate = auction::DELEGATION_RATE_DENOMINATOR;
 const UPDATED_CALL_CONTRACT_COST: Cost = 12_345;
@@ -196,6 +198,7 @@ fn upgraded_add_bid_and_withdraw_bid_have_expected_costs() {
         DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
         DEFAULT_MAX_STORED_VALUE_SIZE,
         DEFAULT_MAX_DELEGATOR_SIZE_LIMIT,
+        DEFAULT_MINIMUM_DELEGATION_AMOUNT,
         WasmConfig::default(),
         new_system_config,
     );
@@ -437,6 +440,7 @@ fn upgraded_delegate_and_undelegate_have_expected_costs() {
         DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
         DEFAULT_MAX_STORED_VALUE_SIZE,
         DEFAULT_MAX_DELEGATOR_SIZE_LIMIT,
+        DEFAULT_MINIMUM_DELEGATION_AMOUNT,
         WasmConfig::default(),
         new_system_config,
     );
@@ -879,6 +883,7 @@ fn should_verify_wasm_add_bid_wasm_cost_is_not_recursive() {
         DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
         DEFAULT_MAX_STORED_VALUE_SIZE,
         DEFAULT_MAX_DELEGATOR_SIZE_LIMIT,
+        DEFAULT_MINIMUM_DELEGATION_AMOUNT,
         new_wasm_config,
         new_system_config,
     );