+++ node/src/types/chainspec/core_config.rs
@@ -34,6 +34,8 @@ pub struct CoreConfig {
     pub(crate) max_runtime_call_stack_height: u32,
     /// Maximum serialized size of values stored in global state.
     pub(crate) max_stored_value_size: u32,
+    /// The minimum bound of motes that can be delegated to a validator.
+    pub(crate) minimum_delegation_amount: u64,
 }
 
 #[cfg(test)]
@@ -53,6 +55,7 @@ impl CoreConfig {
         let max_associated_keys = rng.gen();
         let max_runtime_call_stack_height = rng.gen();
         let max_stored_value_size = rng.gen();
+        let minimum_delegation_amount = rng.gen::<u32>() as u64;
 
         CoreConfig {
             era_duration,
@@ -65,6 +68,7 @@ impl CoreConfig {
             max_associated_keys,
             max_runtime_call_stack_height,
             max_stored_value_size,
+            minimum_delegation_amount,
         }
     }
 }
@@ -82,6 +86,7 @@ impl ToBytes for CoreConfig {
         buffer.extend(self.max_associated_keys.to_bytes()?);
         buffer.extend(self.max_runtime_call_stack_height.to_bytes()?);
         buffer.extend(self.max_stored_value_size.to_bytes()?);
+        buffer.extend(self.minimum_delegation_amount.to_bytes()?);
         Ok(buffer)
     }
 
@@ -96,6 +101,7 @@ impl ToBytes for CoreConfig {
             + self.max_associated_keys.serialized_length()
             + self.max_runtime_call_stack_height.serialized_length()
             + self.max_stored_value_size.serialized_length()
+            + self.minimum_delegation_amount.serialized_length()
     }
 }
 
@@ -111,6 +117,7 @@ impl FromBytes for CoreConfig {
         let (max_associated_keys, remainder) = u32::from_bytes(remainder)?;
         let (max_runtime_call_stack_height, remainder) = u32::from_bytes(remainder)?;
         let (max_stored_value_size, remainder) = u32::from_bytes(remainder)?;
+        let (minimum_delegation_amount, remainder) = u64::from_bytes(remainder)?;
         let config = CoreConfig {
             era_duration,
             minimum_era_height,
@@ -122,6 +129,7 @@ impl FromBytes for CoreConfig {
             max_associated_keys,
             max_runtime_call_stack_height,
             max_stored_value_size,
+            minimum_delegation_amount,
         };
         Ok((config, remainder))
     }